+++ src/Symfony/Bridge/Doctrine/Tests/Fixtures/BaseUser.php
@@ -2,6 +2,9 @@
 
 namespace Symfony\Bridge\Doctrine\Tests\Fixtures;
 
+use Symfony\Component\Validator\Constraints as Assert;
+use Symfony\Component\Validator\Mapping\ClassMetadata;
+
 /**
  * Class BaseUser.
  */
@@ -45,5 +48,16 @@ class BaseUser
     public function getUsername()
     {
         return $this->username;
+    }
+
+    public static function loadValidatorMetadata(ClassMetadata $metadata): void
+    {
+        $allowEmptyString = property_exists(Assert\Length::class, 'allowEmptyString') ? ['allowEmptyString' => true] : [];
+
+        $metadata->addPropertyConstraint('username', new Assert\Length([
+            'min' => 2,
+            'max' => 120,
+            'groups' => ['Registration'],
+        ] + $allowEmptyString));
     }
 }